token 적용

+++ Woong_Server/src/models/bookmarkModel.js
@@ -1,14 +1,14 @@
 const moment = require('moment')
 
 // 1. 북마크 리스트
-exports.getBookmark = (connection, data) => {
+exports.getBookmark = (connection, user_id) => {
   return new Promise((resolve, reject) => {
     const Query = `select market_id,market_name,title_image 
                    from WP_MARKET 
-                   where market_id in (select market_id from WP_MARKET_BOOKMARK where user_id=?)`
-    connection.query(Query, [data.user_token], (err, result) => {
+                   where market_id 
+                   in (select market_id from WP_MARKET_BOOKMARK where user_id=?)`
+    connection.query(Query, [user_id], (err, result) => {
       err && reject(err)
-      console.log([data.user_id])
       console.log(result)
       console.log(result.length)
       resolve(result)
@@ -17,12 +17,12 @@ exports.getBookmark = (connection, data) => {
 }
 
 // 2. 북마크 등록
-exports.addBookmark = (connection, data) => {
+exports.addBookmark = (connection, market_id, user_id) => {
   return new Promise((resolve, reject) => {
     const Query = `INSERT INTO 
                   WP_MARKET_BOOKMARK(market_id,user_id,cr_dt, cr_user) 
                   VALUES (?,?,?,?)`
-    connection.query(Query, [data.market_id, data.user_token, moment().format('YYYY-MM-DD hh:mm:ss'), data.user_token], (err) => {
+    connection.query(Query, [market_id, user_id, moment().format('YYYY-MM-DD hh:mm:ss'), data.user_id], (err) => {
       err && reject(err)
       resolve({})
     })
@@ -30,10 +30,10 @@ exports.addBookmark = (connection, data) => {
 }
 
 // 3. 북마크 제거
-exports.deleteBookmark = (connection, data) => {
+exports.deleteBookmark = (connection, market_id, user_id) => {
   return new Promise((resolve, reject) => {
-    const Query = 'DELETE FROM WP_MARKET_BOOKMARK WHERE market_id=?'
-    connection.query(Query, [data.market_id], (err) => {
+    const Query = 'DELETE FROM WP_MARKET_BOOKMARK WHERE market_id=? AND user_id=?'
+    connection.query(Query, [market_id, user_id], (err) => {
       err && reject(err)
       resolve({})
     })
@@ -41,7 +41,7 @@ exports.deleteBookmark = (connection, data) => {
 }
 
 // 사용자 북마크 개수
-exports.selectCountByBookmark = (connection, data) => {
+exports.selectCountByBookmark = (connection, user_id, market_id) => {
   console.log('model in')
   return new Promise((resolve, reject) => {
     const Query = `
@@ -52,7 +52,7 @@ exports.selectCountByBookmark = (connection, data) => {
     WHERE
       user_id = ? AND market_id = ?
     `
-    connection.query(Query, [data.user_token, data.market_id], (err, result) => {
+    connection.query(Query, [user_id, market_id], (err, result) => {
       err && reject(new Error(err))
       
       resolve(result)
